+++ src/utils.js
@@ -55,9 +55,16 @@
    * Define a typed enum from an array of string or an associative array of number.
    * Enum items are instances of PseudoEnumItemClass.
    *
-   * var enumExample = that.enum(["Label1", "Label2"], EnumExampleItem = that.EnumItem);
-   * enumExample.Label1 instanceof EnumExampleItem; // true
-   * 
+   * var ArrayEnumItem;
+   * var arrayEnum = that.enum(['Label1', 'Label2'], ArrayEnumItem = that.EnumItem);
+   * arrayEnum.Label2 instanceof ArrayEnumItem; // true
+   * objectEnum.Label2.value === 1; // true
+   *
+   * var ObjectEnumItem;
+   * var objectEnum = that.enum({'Label1': 10, 'Label2': 20}, ObjectEnumItem = that.EnumItem);
+   * objectEnum.Label1 instanceof ObjectEnumItem; // true
+   * objectEnum.Label1.value === 10; // true
+   *
    * @param {(string[]|Object.<string, number>)} dataList - list of keys, with or without
    * @param {function} [PseudoEnumItemClass] - class of enum items (used for type checking)
    */
@@ -74,5 +81,18 @@
       }
     }
     return result;
+  };
+
+  /**
+   * Improved version of typeof operator.
+   * Supports: Undefined, Null, Boolean, Number, String, Array, Function, Object, RegExp, Arguments, Math, JSON,
+   * ImPro, TypedArray(Uint8Clamped, Uint8, Uint16, Uint32, Int8, Int16, Int32, Float32, Float64).
+   * See Jasmine tests for examples.
+   *
+   * @param {*} instance - object or primitive whose type is to be returned
+   * @return {string} Type as a capitalized string
+   */
+  that.typeof = function(instance) {
+    return (instance === that) ? 'ImPro' : Object.prototype.toString.call(instance).slice(8, -1);
   };
 })(ImPro);