+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
-Current version is **0.1.2** *(02.04.2018)*
+Current version is **0.1.3** *(07.04.2018)*
+
+#### Version 0.1.3 (07.04.2018)
+* Add typeof method in utils
+* Complete utils tests
 
 #### Version 0.1.2 (02.04.2018)
 * Gaussian kernel for convolution